Fixed the negative c problem
The negative c values arose because one can reverse the direction of time by a parameter transformation. If the bootstrap-estimators receive the original estimation as a starting guess, then the time direction stays the same. The results show a relatively smooth distribution of the bootstrapped parameters. The standard deviations of the parameters (calculated via the bootstrapped sample) can be used to calculate when the growth of SPC1 starts and ends, resulting in a rather narrow time frame during which most of the cultural evolution takes place.