Try to match polity change to SPC1-growth
This file tells us the relationship between previous polities. I use a very broad distinction between continuous and non-continuity. My goal is to check if the short time interval of SPC1-growth from the other notebook corresponds with a large amount of discontinuities. (The TimeZero.csv is produced by the other notebook, the xlsx was sent to me by Dan). I still need to adjust the polity-change-years by subtracting the TimeZero-year. To do so, I still need to discard all NGAs that never reach a TimeZero-year.