Test Reltime Relevance
Updating the previous notebooks, I export csv-data that helps to test the relevance of RelTime via the same method as Turchin in his 2015 PNAS paper. The test seems to indicate that RelTime does not help us to improve the prediction. However, if we exclude the strange outlier time series with the crisis (Kachi plane?), this might work out instead. But we'll have to test this on another day.